Update patch set 1
Patch Set 1: Code-Review-1

Please add a coding style configuration. I suggest using the AirBnB configuration, which is the one we are using in the other JS components, unless there is another one recommended for React projects. This can be done in this initial commit or if you prefer, in a subsequent one.

This coding style configuration will probably force you to change `var` to `const`, for instance, which would be part of my reviews, but I'll wait the for the tool suggestions before commenting about it.

Patch-set: 1
Reviewer: Thiago Cardoso <1013086@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1